Refactor exception package to message
